+++ blueprints/index_blueprint.py
@@ -1,0 +1,7 @@
+from flask import Blueprint
+
+blueprint = Blueprint('index', __name__)
+
+@blueprint.route('/')
+def index():
+    return 'Welcome to Jukebrox! Web app under development.'
